API Fxing code style errors step8
# What
- Fixing assertions
- Add missing documentation
- mark implementation as internal
- Fix member/method ordering
