+++ Api/src/asserts/DictionaryAssert.cs
@@ -9,7 +9,7 @@ using System.Linq;
 
 using CommandLine;
 
-public sealed class DictionaryAssert<TKey, TValue> : AssertBase<IEnumerable>, IDictionaryAssert<TKey, TValue>
+internal sealed class DictionaryAssert<TKey, TValue> : AssertBase<IEnumerable>, IDictionaryAssert<TKey, TValue>
     where TKey : notnull
 {
     public DictionaryAssert(IDictionary<TKey, TValue>? current)
@@ -166,9 +166,7 @@ public sealed class DictionaryAssert<TKey, TValue> : AssertBase<IEnumerable>, ID
     public new IDictionaryAssert<TKey, TValue> OverrideFailureMessage(string message)
         => (IDictionaryAssert<TKey, TValue>)base.OverrideFailureMessage(message);
 
-#pragma warning disable CA1000 // Do not declare static members on generic types
-    public static DictionaryAssert<TKey, TValue> From(IDictionary? current)
-#pragma warning restore CA1000 // Do not declare static members on generic types
+    internal static DictionaryAssert<TKey, TValue> From(IDictionary? current)
         => new(current);
 
     private TValue? TryGetValue(TKey key)
