+++ Api/src/asserts/ExceptionAssert.cs
@@ -11,7 +11,7 @@ using System.Runtime.ExceptionServices;
 using Core.Execution.Exceptions;
 using Core.Extensions;
 
-public sealed class ExceptionAssert<TException> : IExceptionAssert
+internal sealed class ExceptionAssert<TException> : IExceptionAssert
     where TException : Exception
 {
     public ExceptionAssert(Action action)
@@ -20,7 +20,9 @@ public sealed class ExceptionAssert<TException> : IExceptionAssert
         {
             action.Invoke();
         }
+#pragma warning disable CA1031
         catch (Exception e)
+#pragma warning restore CA1031
         {
             var capturedException = ExceptionDispatchInfo.Capture(e.InnerException ?? e);
             Current = (TException)capturedException.SourceException;
@@ -53,9 +55,7 @@ public sealed class ExceptionAssert<TException> : IExceptionAssert
     {
         int currentLine;
         if (Current is TestFailedException e)
-        {
             currentLine = e.LineNumber;
-        }
         else
         {
             var stackFrame = new StackTrace(Current!, true).GetFrame(0);
@@ -72,9 +72,7 @@ public sealed class ExceptionAssert<TException> : IExceptionAssert
         var fullPath = Path.GetFullPath(fileName);
         string currentFileName;
         if (Current is TestFailedException e)
-        {
             currentFileName = e.FileName ?? string.Empty;
-        }
         else
         {
             var stackFrame = new StackTrace(Current!, true).GetFrame(0);
@@ -109,16 +107,16 @@ public sealed class ExceptionAssert<TException> : IExceptionAssert
         return this;
     }
 
-    private void ThrowTestFailureReport(string message)
-    {
-        var failureMessage = CustomFailureMessage ?? message;
-        throw new TestFailedException(failureMessage);
-    }
-
     internal string? GetExceptionStackTrace()
     {
         if (Current is TestFailedException tfe)
             return tfe.StackTrace;
         return Current?.StackTrace ?? null;
+    }
+
+    private void ThrowTestFailureReport(string message)
+    {
+        var failureMessage = CustomFailureMessage ?? message;
+        throw new TestFailedException(failureMessage);
     }
 }
