+++ Api/src/asserts/NumberAssert.cs
@@ -6,7 +6,7 @@ namespace GdUnit4.Asserts;
 using System;
 using System.Numerics;
 
-public class NumberAssert<TValue> : AssertBase<TValue>, INumberAssert<TValue>
+internal class NumberAssert<TValue> : AssertBase<TValue>, INumberAssert<TValue>
     where TValue : IComparable, IComparable<TValue>, IEquatable<TValue>,
     IAdditionOperators<TValue, TValue, TValue>, ISubtractionOperators<TValue, TValue, TValue>
 {
