+++ stylecop.ruleset
@@ -7,4 +7,7 @@
     <Rule Id="SA1503" Action="None"/> <!-- Disable BracesMustNotBeOmitted -->
     <Rule Id="SA1520" Action="None"/> <!-- Disable The opening and closing braces of a chained Rules-->
   </Rules>
+  <Rules AnalyzerId="Microsoft.CodeAnalysis.CSharp.Features" RuleNamespace="Microsoft.CodeAnalysis.CSharp.Features">
+    <Rule Id="IDE0011" Action="None"/> <!-- Disable Add braces to 'if' statement -->
+  </Rules>
 </RuleSet>
